Cambios agregados para evitar confusiones y programación compleja
