+++ src/transit/Transit.java
@@ -1,18 +1,37 @@
 package transit;
 
+import payment.Discount;
 import ticket.Category;
 import ticket.GovAuthorities;
 import ticket.Ticket;
 
-public class Transit extends Ticket {
+public class Transit extends Ticket implements Discount {
 
     public Category trafficType = Category.TRAFFIC ;
-    private static final int MINIMUM_FINE_TRAFFIC = 1;
-    private static final double MAX_DISCOUNT_TRAFFIC = 0.50;
 
     public Transit(Category trafficType,
                    GovAuthorities authorityName) {
         super(authorityName);
         this.trafficType = trafficType;
+    }
+
+    @Override
+    public double applyDiscounts(long daysSinceIssue, Category category, double fineAmount) {
+
+        if (daysSinceIssue <= Discount.MAJOR_DISCOUNT) {
+            double percentage = (1 - Category.TRAFFIC.maxDiscount);
+            System.out.println("Tienes un descuento del " + percentage + "%. Total a pagar: ");
+            return fineAmount * percentage;
+
+        } else if (daysSinceIssue <= Discount.MINOR_DISCOUNT) {
+            double percentage = (1 - Category.TRAFFIC.maxDiscount / 3);
+            System.out.println("Tienes un descuento del " + percentage + "%. Total a pagar: ");
+            return fineAmount * percentage;
+
+        } else if (daysSinceIssue <= Discount.DEADLINE) {
+            System.out.println("Descuentos no disponibles. Total a pagar: ");
+            return fineAmount;
+        }
+        throw new IllegalStateException("Días desde la emisión inválidos o categoría no manejada.");
     }
 }
